Swept out all crap previously made
