+++ src/media/buffers/frame-buffer.js
@@ -69,15 +69,6 @@ export class FrameBuffer {
     return frame;
   }
 
-  popFirstFrame() {
-    if (this._frames.isEmpty()) {
-      return null;
-    }    
-    const frame = this._frames.pop();
-    this._updateFirstFrameTs();
-    return frame;
-  }
-
   absorb(frameBuffer) {
     frameBuffer.forEach((frame) => {
       if (frame.timestamp > this._lastFrameTs) {
