+++ src/state-manager.js
@@ -1,6 +1,5 @@
 import { STATE, IDX } from "./shared/values";
 import { isSharedBuffer } from "./shared/shared-buffer";
-import { LoggersFactory } from "@/shared/logger";
 
 const U32POWER = 0x0100000000;
 
@@ -15,14 +14,11 @@ export class StateManager {
       options.shared ?? (typeof Atomics !== "undefined" && isSharedBuffer(sab));
     this._suppressNotify = false;
     this._port = null;
-    this.displayName = options.name ?? "Untitled";
-    this._logger = LoggersFactory.create("nimio", "State" + this.displayName);
     this._onPortMessage = this._handlePortMessage.bind(this);
     this._sendInit = options.sendInit ?? true;
     if (options.port) {
-      this.attachPort(options.port) ;
+      this.attachPort(options.port);
     }
-    console.log("Create StateManager", this.displayName);
   }
 
   get value() {
@@ -34,22 +30,18 @@ export class StateManager {
   }
 
   isStopped() {
-    this._logger.debug(`${this.displayName} state = ${this.value}`)
     return this.value === STATE.STOPPED;
   }
 
   isPlaying() {
-    this._logger.debug(`${this.displayName} state = ${this.value}`)
     return this.value === STATE.PLAYING;
   }
 
   isPaused() {
-    this._logger.debug(`${this.displayName} state = ${this.value}`)
     return this.value === STATE.PAUSED;
   }
 
   start() {
-    this._logger.debug(`${this.displayName} state := PLAYING (${STATE.PLAYING})`)
     this.value = STATE.PLAYING;
   }
 
@@ -70,18 +62,14 @@ export class StateManager {
   }
 
   getCurrentTsSmp() {
-    const smpCnt = this._atomicLoad64(IDX.CURRENT_TS);
-    this._logger.debug(`${this.displayName} getCurrentTsSmp ${smpCnt}`)
-    return smpCnt;
+    return this._atomicLoad64(IDX.CURRENT_TS);
   }
 
   setCurrentTsSmp(smpCnt) {
-    this._logger.debug(`${this.displayName} setCurrentTsSmp ${smpCnt}`)
     this._atomicStore64(IDX.CURRENT_TS, smpCnt);
   }
 
   incCurrentTsSmp(smpCnt) {
-    this._logger.debug(`${this.displayName} incCurrentTsSmp ${smpCnt}`)
     return this._atomicAdd64(IDX.CURRENT_TS, smpCnt);
   }
 
@@ -106,13 +94,10 @@ export class StateManager {
   }
 
   getPlaybackStartTsUs() {
-    const tsUs = this._atomicLoad64(IDX.PLAYBACK_START_TS);
-    this._logger.debug(`${this.displayName} PlaybackStartTsUs == ${tsUs}`)
-    return tsUs;
+    return this._atomicLoad64(IDX.PLAYBACK_START_TS);
   }
 
   setPlaybackStartTsUs(tsUs) {
-    this._logger.debug(`${this.displayName} PlaybackStartTsUs := ${tsUs}`)
     this._atomicStore64(IDX.PLAYBACK_START_TS, tsUs);
   }
 
@@ -291,7 +276,6 @@ export class StateManager {
     if (this._port) {
       this._detachPort();
     }
-    this._logger.debug(`${this.displayName} attachPort`)
     this._port = port;
     this._port.addEventListener("message", this._onPortMessage);
     if (this._sendInit) {
@@ -325,36 +309,20 @@ export class StateManager {
     }
   }
 
-  _notify(op, idx, value, skipWorker) {
-    const skipSend = (this._shared || !this._port || this._suppressNotify)
-    if (idx == 9 || idx == 10) { //PLAYBACK_START_TS 
-      if (skipSend) {
-        this._logger.debug(`State ${this.displayName} send ${op} PLAYBACK_START_TS SKIP ${!!this._shared} ${!!this._port} ${this._suppressNotify}`)
-      } else {
-        this._logger.debug(`State ${this.displayName} send ${op} PLAYBACK_START_TS ${value}.`)
-      }
-    }
-    const msg = {
+  _notify(op, idx, value) {
+    if (this._shared || !this._port || this._suppressNotify) return;
+    this._port.postMessage({
       type: "state:update",
       op,
       idx,
       value,
-    };
-    if (this.onWorkerClient && !skipWorker) this.onWorkerClient(msg);
-    if (skipSend) return;
-    this._port.postMessage(msg);
-
+    });
   }
 
   _handlePortMessage(ev) {
     const msg = ev.data;
     if (!msg || msg.type !== "state:update") return;
-   if (msg.idx == 9 || msg.idx == 10) { //PLAYBACK_START_TS 
-      const smpCnt = this._atomicLoad64(IDX.PLAYBACK_START_TS);      
-      this._logger.debug(`State ${this.displayName} received ${msg.op} PLAYBACK_START_TS ${msg.value} to ${smpCnt}`)
-  //  } else {
-  //    this._logger.debug(`State ${this.displayName} received ${msg.op}`)
-   } 
+
     this._suppressNotify = true;
     if (msg.op === "init" && Array.isArray(msg.value)) {
       this._flags.set(msg.value);
@@ -368,21 +336,10 @@ export class StateManager {
       this._atomicAdd64([msg.idx, msg.idx + 1], msg.value);
     }
     this._suppressNotify = false;
-    if (this.onWorkerClient) {
-      const resendMsg = {
-        type: "state:update",
-        op: msg.op,
-        idx: msg.idx,
-        value: msg.value,
-      };
-      this.onWorkerClient(resendMsg);
-    }
   }
 
   _detachPort() {
     if (!this._port) return;
-    this._logger.debug(`${this.displayName} detachPort`)
-
     this._port.removeEventListener("message", this._onPortMessage);
     this._port = null;
   }
